fix: merge SubscriptionsChecker into SubscriptionManager
We will use a single database connection and keep all subscription
management in a single place
